requirements.txt
# required
fiona>=1.8.19
pandas>=1.1.0
pyproj>=3.0.1
shapely>=1.7.1
packaging
geopandas>=0.13.0
dask
dask-geopandas

# geodatabase access
psycopg2-binary>=2.8.0
SQLAlchemy>=1.3
sshtunnel

# geocoding
geopy

# plotting
matplotlib>=3.3.4
mapclassify
seaborn

# PostGIS writing
GeoAlchemy2
